refactor: geocode directory
